Change from file size to validation for router.db
This rewrites parts of update_routerdb to be more flexible. It should
run on "sh" instead of bash, which is useful for my friends running
FreeBSD without an installed copy of bash.

The validation has been updated to parse the file and try to match it
with what is expected of router.db. This is better than just checking
the amount of lines because some networks are bigger than others, and
HTML error messages might end up being 100 lines long.

The path for the router.db file is now an environment variable so it can
be overridden. You can also specify it on the commandline.

Finally, the /etc/netbox_auth_token file is optional.